refactor(frontend): extract ModifierGroupAssignments + AllergenDietaryFields
Dissolves the bulk of the ~717-line ProductEdit/ComboProductEdit clone by
pulling the genuinely-identical pieces into shared components.

- ModifierGroupAssignments: the modifier-group assignment block (state, data
  hooks, 5 handlers, ordering UI) was duplicated verbatim (with drift) in both
  product editors. One component now owns it; both pages render
  <ModifierGroupAssignments brandSlug productId />. Unifies the drifted move
  handlers and the group name/count resolution.
- AllergenDietaryFields: one component, two modes. Editable (Controller-bound)
  for simple products; read-only for combos, showing a derived roll-up —
  allergens = union, dietary tags = intersection of the selected component
  products (data already on the client via ProductListItem). Combos previously
  showed no allergen/dietary info at all.

No change to simple-product behavior. fallow duplication 19.2% -> 17.5%
(-571 duplicated lines, 5 fewer clone groups). type-check, lint and 343 unit
tests all green.
