+++ src/frontend/src/features/admin/pages/ComboProductEdit.tsx
@@ -1,13 +1,8 @@
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
 import { Controller } from 'react-hook-form';
 import { useProducts, useDeleteProduct } from '../hooks/useProducts';
-import {
-  useProductModifierGroups,
-  useModifierGroups,
-  useSetProductModifierGroups,
-} from '../hooks/useModifierGroups';
 import { useResourceForm } from '../forms/useResourceForm';
 import { productKeys } from '../hooks/useProducts';
 import { productsApi } from '../../../api/products';
@@ -15,12 +10,14 @@ import { comboProductEditSchema, type ComboProductEditFormValues } from './schem
 import type {
   SupportedLocale,
   ProductListItem,
-  ProductModifierGroupResponse,
   Product,
 } from '../../../types/common';
+import { Allergen, DietaryTag, ALLERGEN_KEYS, DIETARY_TAG_KEYS } from '../../../types/common';
 import type { UpdateComboProductRequest } from '../../../api/products';
 import { labelStyle, inputStyle, secondaryButtonStyle, RequiredMark, FieldError } from '../forms/adminFormStyles';
 import { ComboTranslationFields } from '../forms/ComboTranslationFields';
+import { ModifierGroupAssignments } from '../forms/ModifierGroupAssignments';
+import { AllergenDietaryFields } from '../forms/AllergenDietaryFields';
 
 // ---------------------------------------------------------------------------
 // Constants
@@ -71,22 +68,6 @@ export function ComboProductEdit() {
 
   const { data: allProducts } = useProducts(resolvedBrandSlug);
   const deleteProduct = useDeleteProduct(resolvedBrandSlug);
-
-  // Modifier groups (kept imperative — separate resource / mutation)
-  const { data: productModifierGroups } = useProductModifierGroups(resolvedBrandSlug, resolvedProductId);
-  const { data: allModifierGroups } = useModifierGroups(resolvedBrandSlug);
-  const setProductModifierGroups = useSetProductModifierGroups(resolvedBrandSlug, resolvedProductId);
-  const [assignedGroups, setAssignedGroups] = useState<ProductModifierGroupResponse[]>([]);
-  const [assignmentsInitialized, setAssignmentsInitialized] = useState(false);
-  const [selectedGroupToAdd, setSelectedGroupToAdd] = useState('');
-
-  // Populate assigned modifier groups when data arrives
-  useEffect(() => {
-    if (productModifierGroups !== undefined && !assignmentsInitialized) {
-      setAssignedGroups(productModifierGroups);
-      setAssignmentsInitialized(true);
-    }
-  }, [productModifierGroups, assignmentsInitialized]);
 
   // Active translation tab (UI-only state — not part of the form schema)
   const [activeTab, setActiveTab] = useState<SupportedLocale>('nl');
@@ -194,65 +175,6 @@ export function ComboProductEdit() {
   }
 
   // ---------------------------------------------------------------------------
-  // Modifier group assignment handlers (imperative — separate resource)
-  // ---------------------------------------------------------------------------
-
-  function handleAddModifierGroup() {
-    if (!selectedGroupToAdd) return;
-    const alreadyAssigned = assignedGroups.some((g) => g.modifierGroupId === selectedGroupToAdd);
-    if (alreadyAssigned) return;
-
-    const groupInfo = allModifierGroups?.find((g) => g.id === selectedGroupToAdd);
-    if (!groupInfo) return;
-
-    const newGroup: ProductModifierGroupResponse = {
-      modifierGroupId: groupInfo.id,
-      name: groupInfo.name,
-      sortOrder: assignedGroups.length,
-      modifiers: [],
-    };
-    setAssignedGroups((prev) => [...prev, newGroup]);
-    setSelectedGroupToAdd('');
-  }
-
-  function handleRemoveAssignedGroup(modifierGroupId: string) {
-    setAssignedGroups((prev) =>
-      prev
-        .filter((g) => g.modifierGroupId !== modifierGroupId)
-        .map((g, index) => ({ ...g, sortOrder: index })),
-    );
-  }
-
-  function handleMoveGroupUp(index: number) {
-    if (index === 0) return;
-    setAssignedGroups((prev) => {
-      const next = [...prev];
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion -- index is in-bounds (index > 0) and next is a copy of prev
-      [next[index - 1], next[index]] = [next[index]!, next[index - 1]!];
-      return next.map((g, i) => ({ ...g, sortOrder: i }));
-    });
-  }
-
-  function handleMoveGroupDown(index: number) {
-    setAssignedGroups((prev) => {
-      if (index >= prev.length - 1) return prev;
-      const next = [...prev];
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion -- index and index+1 are in-bounds (index < length-1) and next is a copy of prev
-      [next[index], next[index + 1]] = [next[index + 1]!, next[index]!];
-      return next.map((g, i) => ({ ...g, sortOrder: i }));
-    });
-  }
-
-  function handleSaveAssignments() {
-    setProductModifierGroups.mutate({
-      assignments: assignedGroups.map((g) => ({
-        modifierGroupId: g.modifierGroupId,
-        sortOrder: g.sortOrder,
-      })),
-    });
-  }
-
-  // ---------------------------------------------------------------------------
   // Loading / error states
   // ---------------------------------------------------------------------------
 
@@ -292,6 +214,21 @@ export function ComboProductEdit() {
   // Pre-compute error messages to avoid complex type inference inside JSX
   const nlNameError = (errors.translations?.nl?.name as { message?: string } | undefined)?.message;
   const componentIdsError = (errors.componentProductIds as { message?: string } | undefined)?.message;
+
+  // Allergens/dietary tags are derived (read-only) for combos: a combo carries
+  // the union of its components' allergens and the intersection of their dietary tags.
+  const selectedComponents = watch('componentProductIds')
+    .map((id) => allProducts?.find((p) => p.id === id))
+    .filter((p): p is ProductListItem => p !== undefined);
+  const comboAllergens = ALLERGEN_KEYS.map((key) => Allergen[key]).filter((value) =>
+    selectedComponents.some((p) => p.allergens.includes(value)),
+  );
+  const comboDietaryTags =
+    selectedComponents.length === 0
+      ? []
+      : DIETARY_TAG_KEYS.map((key) => DietaryTag[key]).filter((value) =>
+          selectedComponents.every((p) => p.dietaryTags.includes(value)),
+        );
 
   return (
     <main style={{ padding: '1.5rem', maxWidth: '40rem' }}>
@@ -542,6 +479,13 @@ export function ComboProductEdit() {
           />
         </section>
 
+        {/* Allergens & dietary tags — derived (read-only) from the selected components */}
+        <AllergenDietaryFields
+          mode="readonly"
+          allergens={comboAllergens}
+          dietaryTags={comboDietaryTags}
+        />
+
         {/* API error */}
         {submitError != null && (
           <p style={{ color: '#dc2626', marginBottom: '1rem', fontSize: '0.875rem' }}>
@@ -591,167 +535,8 @@ export function ComboProductEdit() {
         </div>
       </form>
 
-      {/* Modifier Groups Section */}
-      <section
-        style={{
-          marginTop: '2.5rem',
-          borderTop: '2px solid #e5e7eb',
-          paddingTop: '1.5rem',
-        }}
-      >
-        <h2 style={{ fontSize: '1.125rem', fontWeight: 700, marginBottom: '1rem' }}>
-          {t('admin.modifierGroups.title')}
-        </h2>
-
-        {/* Assigned groups list */}
-        {assignedGroups.length === 0 ? (
-          <p style={{ color: '#6b7280', marginBottom: '1rem' }}>
-            {t('admin.modifierGroups.noAssignedGroups')}
-          </p>
-        ) : (
-          <div style={{ marginBottom: '1rem' }}>
-            {assignedGroups.map((group, index) => (
-              <div
-                key={group.modifierGroupId}
-                style={{
-                  display: 'flex',
-                  alignItems: 'center',
-                  gap: '0.5rem',
-                  padding: '0.625rem 0.75rem',
-                  border: '1px solid #e5e7eb',
-                  borderRadius: '0.375rem',
-                  marginBottom: '0.5rem',
-                  background: '#f9fafb',
-                }}
-              >
-                <div style={{ flex: 1 }}>
-                  <span style={{ fontWeight: 600, fontSize: '0.9rem' }}>{group.name}</span>
-                  {group.modifiers.length > 0 && (
-                    <span style={{ color: '#6b7280', fontSize: '0.75rem', marginLeft: '0.5rem' }}>
-                      {group.modifiers.length} {t('admin.modifierGroups.modifiers').toLowerCase()}
-                    </span>
-                  )}
-                </div>
-                <button
-                  type="button"
-                  onClick={() => { handleMoveGroupUp(index); }}
-                  disabled={index === 0}
-                  style={{
-                    ...reorderButtonStyle,
-                    opacity: index === 0 ? 0.3 : 1,
-                    cursor: index === 0 ? 'not-allowed' : 'pointer',
-                  }}
-                >
-                  &#9650;
-                </button>
-                <button
-                  type="button"
-                  onClick={() => { handleMoveGroupDown(index); }}
-                  disabled={index === assignedGroups.length - 1}
-                  style={{
-                    ...reorderButtonStyle,
-                    opacity: index === assignedGroups.length - 1 ? 0.3 : 1,
-                    cursor: index === assignedGroups.length - 1 ? 'not-allowed' : 'pointer',
-                  }}
-                >
-                  &#9660;
-                </button>
-                <button
-                  type="button"
-                  onClick={() => { handleRemoveAssignedGroup(group.modifierGroupId); }}
-                  style={{
-                    padding: '0.125rem 0.5rem',
-                    fontSize: '0.75rem',
-                    background: '#fff',
-                    border: '1px solid #fca5a5',
-                    borderRadius: '0.25rem',
-                    color: '#dc2626',
-                    cursor: 'pointer',
-                  }}
-                >
-                  {t('admin.modifierGroups.removeModifier')}
-                </button>
-              </div>
-            ))}
-          </div>
-        )}
-
-        {/* Add group dropdown */}
-        {allModifierGroups !== undefined && allModifierGroups.length > 0 && (
-          <div style={{ display: 'flex', gap: '0.5rem', marginBottom: '1rem', alignItems: 'center' }}>
-            <select
-              value={selectedGroupToAdd}
-              onChange={(e) => { setSelectedGroupToAdd(e.target.value); }}
-              style={{
-                padding: '0.5rem 0.75rem',
-                border: '1px solid #d1d5db',
-                borderRadius: '0.375rem',
-                fontSize: '0.9rem',
-                flex: 1,
-                maxWidth: '20rem',
-              }}
-            >
-              <option value="">{t('admin.comboProducts.selectModifierGroup')}</option>
-              {allModifierGroups
-                .filter((g) => !assignedGroups.some((a) => a.modifierGroupId === g.id))
-                .map((g) => (
-                  <option key={g.id} value={g.id}>
-                    {g.name}
-                  </option>
-                ))}
-            </select>
-            <button
-              type="button"
-              onClick={handleAddModifierGroup}
-              disabled={!selectedGroupToAdd}
-              style={{
-                padding: '0.5rem 1rem',
-                background: '#f9fafb',
-                border: '1px solid #d1d5db',
-                borderRadius: '0.375rem',
-                cursor: selectedGroupToAdd ? 'pointer' : 'not-allowed',
-                opacity: selectedGroupToAdd ? 1 : 0.5,
-                fontSize: '0.875rem',
-              }}
-            >
-              + Add
-            </button>
-          </div>
-        )}
-
-        {/* Save assignments */}
-        {setProductModifierGroups.isError && (
-          <p style={{ color: '#dc2626', marginBottom: '0.75rem', fontSize: '0.875rem' }}>
-            {setProductModifierGroups.error instanceof Error
-              ? setProductModifierGroups.error.message
-              : 'Failed to save assignments. Please try again.'}
-          </p>
-        )}
-
-        {setProductModifierGroups.isSuccess && (
-          <p style={{ color: '#16a34a', marginBottom: '0.75rem', fontSize: '0.875rem' }}>
-            Assignments saved.
-          </p>
-        )}
-
-        <button
-          type="button"
-          onClick={handleSaveAssignments}
-          disabled={setProductModifierGroups.isPending}
-          style={{
-            padding: '0.5rem 1.25rem',
-            background: '#111827',
-            color: '#fff',
-            border: 'none',
-            borderRadius: '0.375rem',
-            cursor: setProductModifierGroups.isPending ? 'not-allowed' : 'pointer',
-            fontWeight: 600,
-            opacity: setProductModifierGroups.isPending ? 0.6 : 1,
-          }}
-        >
-          {setProductModifierGroups.isPending ? 'Saving...' : t('admin.modifierGroups.saveAssignments')}
-        </button>
-      </section>
+      {/* Modifier Groups */}
+      <ModifierGroupAssignments brandSlug={resolvedBrandSlug} productId={resolvedProductId} />
     </main>
   );
 }
